Changes to compile under gcc more smoothly (observe the ugly constructs required by gcc in MuxRenderer3D)
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@156 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ misc/BezierDataSource.cpp
@@ -242,7 +242,10 @@ void misc::BezierDataSource::loadBezDat(const vislib::TString& filename) {
             int i2 = vislib::CharTraitsA::ParseInt(bezDat[i].Word(2));
             int i3 = vislib::CharTraitsA::ParseInt(bezDat[i].Word(3));
             int i4 = vislib::CharTraitsA::ParseInt(bezDat[i].Word(4));
-            if ((i1 >= pcnt) || (i2 >= pcnt) || (i3 >= pcnt) || (i4 >= pcnt)) {
+            if ((i1 >= static_cast<int>(pcnt))
+                    || (i2 >= static_cast<int>(pcnt))
+                    || (i3 >= static_cast<int>(pcnt))
+                    || (i4 >= static_cast<int>(pcnt))) {
                 throw vislib::Exception("Point index out of range", __FILE__, __LINE__);
             }
             this->curves.Add(vislib::math::BezierCurve<misc::BezierDataCall::BezierPoint, 3>());
